+++ packages/terra-clinical-item-display/tests/jest/ItemDisplay.test.jsx
@@ -88,7 +88,8 @@ it('should render a display with the secondary textStyle on text', () => {
 it('should render a display with the strikeThrough textStyle on text and default textStyleMeaning provided', () => {
   const itemDisplay = shallowWithIntl(<Display text="Display" textStyle="strikeThrough" />).shallow();
   expect(itemDisplay.contains(<span>Display</span>)).toBe(true);
-  expect(itemDisplay.find('span').first().prop('aria-label')).toEqual('Terra.item-display.textStyleMeaningStrikethrough, Display, Terra.item-display.textStyleMeaningStrikethroughEnd');
+  expect(itemDisplay.find('VisuallyHiddenText').at(0).prop('text')).toEqual('Terra.item-display.textStyleMeaningStrikethrough');
+  expect(itemDisplay.find('VisuallyHiddenText').at(1).prop('text')).toEqual('Terra.item-display.textStyleMeaningStrikethroughEnd');
   expect(itemDisplay.find('.text').prop('className')).toContain('strike-through');
   expect(itemDisplay).toMatchSnapshot();
 });
@@ -96,7 +97,8 @@ it('should render a display with the strikeThrough textStyle on text and default
 it('should render a display with the strikeThrough textStyle on text and aria label based on provided textStyleMeaning prop', () => {
   const itemDisplay = shallowWithIntl(<Display text="Display" textStyle="strikeThrough" textStyleMeaning="Update" />).shallow();
   expect(itemDisplay.contains(<span>Display</span>)).toBe(true);
-  expect(itemDisplay.find('span').first().prop('aria-label')).toEqual('Update, Display, Terra.item-display.textStyleMeaningEnd');
+  expect(itemDisplay.find('VisuallyHiddenText').at(0).prop('text')).toEqual('Update');
+  expect(itemDisplay.find('VisuallyHiddenText').at(1).prop('text')).toEqual('Terra.item-display.textStyleMeaningEnd');
   expect(itemDisplay.find('.text').prop('className')).toContain('strike-through');
   expect(itemDisplay).toMatchSnapshot();
 });
@@ -104,7 +106,8 @@ it('should render a display with the strikeThrough textStyle on text and aria la
 it('should render a display with a non-strikeThrough textStyle on text and aria label based on provided textStyleMeaning prop', () => {
   const itemDisplay = shallowWithIntl(<Display text="Display" textStyle="attention" textStyleMeaning="Important" />).shallow();
   expect(itemDisplay.contains(<span>Display</span>)).toBe(true);
-  expect(itemDisplay.find('span').first().prop('aria-label')).toEqual('Important, Display, Terra.item-display.textStyleMeaningEnd');
+  expect(itemDisplay.find('VisuallyHiddenText').at(0).prop('text')).toEqual('Important');
+  expect(itemDisplay.find('VisuallyHiddenText').at(1).prop('text')).toEqual('Terra.item-display.textStyleMeaningEnd');
   expect(itemDisplay.find('.text').prop('className')).toContain('attention');
   expect(itemDisplay).toMatchSnapshot();
 });