+++ packages/terra-clinical-item-display/CHANGELOG.md
@@ -9,6 +9,9 @@
   * `strong` tag around strong style text.
   * `span` tag around non-strong style text.
 
+* Changed
+  * Improved screen reader support for text styling by utilizing terra-visually-hidden-text.
+
 ## 4.7.0 - (May 15, 2023)
 
 * Deprecated